+++ packages/react/src/Card/Card.module.css.d.ts
@@ -1,5 +1,7 @@
 declare const styles: {
   readonly "Card": string;
+  readonly "Card--fullWidth": string;
+  readonly "Card--variant-minimal": string;
   readonly "Card--colorMode-light": string;
   readonly "Card__link": string;
   readonly "Card--disableAnimation": string;
@@ -8,6 +10,8 @@ declare const styles: {
   readonly "Card--icon": string;
   readonly "Card__label": string;
   readonly "Card__icon": string;
+  readonly "Card--variant-default": string;
+  readonly "Card--colorMode-dark": string;
   readonly "Card__image": string;
   readonly "Card__icon--badge": string;
   readonly "Card__heading": string;