+++ packages/react/src/Image/Image.module.css.d.ts
@@ -6,6 +6,11 @@ declare const styles: {
   readonly "Image--aspect-ratio-16-9": string;
   readonly "Image--aspect-ratio-16-10": string;
   readonly "Image--aspect-ratio-4-3": string;
+  readonly "Image--borderRadius-small": string;
+  readonly "Image--borderRadius-medium": string;
+  readonly "Image--borderRadius-large": string;
+  readonly "Image--borderRadius-xlarge": string;
+  readonly "Image--borderRadius-full": string;
 };
 export = styles;