+++ packages/react/src/Prose/Prose.module.css.d.ts
@@ -1,6 +1,8 @@
 declare const styles: {
   readonly "Prose": string;
+  readonly "Prose--editorial": string;
   readonly "Prose--lineLength": string;
+  readonly "Prose--default": string;
 };
 export = styles;