+++ packages/react/src/global.d.ts
@@ -4,4 +4,5 @@ declare module '*.module.css' {
 }
 
 declare module '*.png'
+declare module '*.webp'
 declare module '*.mp4'