+++ packages/react/src/recipes/seo/Article/Article.module.css.d.ts
@@ -1,0 +1,34 @@
+declare const styles: {
+  readonly "articleCol": string;
+  readonly "asideCol": string;
+  readonly "socialLinks": string;
+  readonly "hero": string;
+  readonly "lightHero": string;
+  readonly "darkHero": string;
+  readonly "parallax": string;
+  readonly "background": string;
+  readonly "foreground": string;
+  readonly "heroImage": string;
+  readonly "heroImageOverlay": string;
+  readonly "heroImageOverlayDark": string;
+  readonly "foregroundContents": string;
+  readonly "articleContents": string;
+  readonly "parallax-foreground": string;
+  readonly "parallax-background": string;
+  readonly "heroLine": string;
+  readonly "heroLineAnimated": string;
+  readonly "wipe": string;
+  readonly "lede": string;
+  readonly "asideHeading": string;
+  readonly "aside": string;
+  readonly "asideContent": string;
+  readonly "asideContent--visible": string;
+  readonly "asideContent--open": string;
+  readonly "tableOfContentsList": string;
+  readonly "tableOfContentsMenuToggle": string;
+  readonly "tableOfContentsNav": string;
+  readonly "tableOfContentsFeaturesBox": string;
+  readonly "tableOfContentsNav--visible": string;
+};
+export = styles;
+